requirements.txt
Django>=3.1.1,<3.2.0
djangorestframework>=3.11.1,<3.12.0
psycopg2-binary>=2.8.6,<2.9.0
dj-rest-auth>=1.1.1,<1.2.0
django-allauth>=0.42,<0.43
requests>=2.24.0,<2.25.0
djangorestframework-simplejwt>=4.4.0,<4.5.0

flake8>=3.8.3,<3.9.0
